requirements.txt
argcomplete==1.12.2
asgiref==3.2.10
awscli==1.19.29
boto3==1.17.27
botocore==1.20.29
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
demjson==2.2.4
Django==3.1.2
django-environ==0.4.5
django-filter==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-storages==1.11.1
djangorestframework==3.12.1
docutils==0.15.2
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
Pillow==7.2.0
pip-tools==6.0.0
placebo==0.9.0
pyasn1==0.4.8
PyMySQL==1.0.2
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.4
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.59.0
troposphere==2.6.4
tyadmin-api-cli==0.8.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
